TelClient/requirements.txt
openai==0.28
pyTelegramBotAPI
python-dotenv